1  新增字段加密设置 ，只适用于  insert   和 find 2  添加了 自动创建索引功能 3 部分代码测试 4  发布0.9.18

+++ mars-core/src/main/java/com/whaleal/mars/monitor/PageFaultsMetrics.java
@@ -20,8 +20,8 @@ public class PageFaultsMetrics extends AbstractMonitor{
         return getOpLatenciesData("page_faults");
     }
 
-    private Long getOpLatenciesData(String key) {
+    private Integer getOpLatenciesData(String key) {
 
-        return serverStatus.get("extra_info",Document.class).get(key,Long.class);
+        return serverStatus.get("extra_info",Document.class).get(key,Integer.class);
     }
 }
